refactor : dto 클래스에서 사용하지 않는 메서드 제거
- 사용하지 않는 메서드들 제거
- 롬복의 Getter와 Setter를 사용하도록 수정
