+++ src/main/java/com/woory/backend/dto/CommentDto.java
@@ -8,7 +8,6 @@ import lombok.NoArgsConstructor;
 import java.util.Date;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.woory.backend.entity.Comment;
 
 @Data
 @NoArgsConstructor
@@ -25,15 +24,4 @@ public class CommentDto {
     private String status;
     private Date commentDate;
 
-
-    public static CommentDto fromComment(Comment comment){
-        return CommentDto.builder()
-            .commentId(comment.getCommentId())
-            .contentId(comment.getContent().getContentId())
-            .userId(comment.getUsers().getUserId())
-            .parentCommentId(comment.getParentComment() != null ? comment.getParentComment().getCommentId() : null)
-            .commentText(comment.getCommentText())
-            .commentDate(comment.getCommentDate())
-            .build();
-    }
 }
