+++ src/main/java/com/woory/backend/dto/GroupInfoDto.java
@@ -30,11 +30,6 @@ public class GroupInfoDto {
 		this.status = groupStatus;
 	}
 
-	public static Map<GroupStatus, List<GroupInfoDto>> toSeparatedByStatus(List<GroupInfoDto> groupInfoList) {
-		return groupInfoList.stream()
-			.collect(Collectors.groupingBy(GroupInfoDto::getStatus));
-	}
-
 	public static GroupInfoDto fromGroup(Group group) {
 		return GroupInfoDto.builder()
 			.groupId(group.getGroupId())
